shell-ui: Fix the test by abstracting the init of DOMRect to setUpTests
